efi: Change pointer integer casting to uintptr_t
Change pointer integer casting from unsigned long to uintptr_t
to improve the code portibility on the architecture which the
pointer is not an unsigned long integer (e.g. CHERI).

+++ include/efi.h
@@ -429,7 +429,7 @@ struct efi_entry_systable {
 static inline struct efi_mem_desc *efi_get_next_mem_desc(
 		struct efi_mem_desc *desc, int desc_size)
 {
-	return (struct efi_mem_desc *)((ulong)desc + desc_size);
+	return (struct efi_mem_desc *)((uintptr_t)desc + desc_size);
 }
 
 /**
